SP-219 refactor: 지원공고 id를 Long 타입으로 변경

+++ src/main/java/com/ludo/study/studymatchingplatform/study/controller/RecruitmentController.java
@@ -30,7 +30,7 @@ public class RecruitmentController {
 
 	@GetMapping
 	public ResponseEntity<RecruitmentPreviewResponses> readRecruitments(
-		@RequestParam(required = false) Integer after, @RequestParam Integer count
+		@RequestParam(required = false) Long after, @RequestParam Integer count
 	) {
 		List<RecruitmentPreviewResponse> recruitments = recruitmentsFindService.findRecruitments(after, count);
